Split ProbeUtils enum & utility functionality
