+++ hazelcast/src/main/java/com/hazelcast/internal/metrics/impl/MethodProbe.java
@@ -16,7 +16,7 @@
 
 package com.hazelcast.internal.metrics.impl;
 
-import static com.hazelcast.internal.metrics.impl.ProbeUtils.getType;
+import static com.hazelcast.internal.metrics.impl.ProbeType.getType;
 import static java.lang.String.format;
 
 import com.hazelcast.internal.metrics.DoubleProbeFunction;
@@ -85,12 +85,12 @@ abstract class MethodProbe implements ProbeFunction {
 
     final boolean isMethodStatic;
     final CachedProbe probe;
-    final ProbeUtils type;
+    final ProbeType type;
     final SourceMetadata sourceMetadata;
     final String probeName;
 
     @SuppressWarnings("checkstyle:executablestatementcount")
-    MethodProbe(final Method method, final Probe probe, final ProbeUtils type, final SourceMetadata sourceMetadata) {
+    MethodProbe(final Method method, final Probe probe, final ProbeType type, final SourceMetadata sourceMetadata) {
         try {
             method.setAccessible(true);
             final MethodHandle unreflected = LOOKUP.unreflect(method);
@@ -162,7 +162,7 @@ abstract class MethodProbe implements ProbeFunction {
     }
 
     static <S> MethodProbe createMethodProbe(final Method method, final Probe probe, final SourceMetadata sourceMetadata) {
-        final ProbeUtils type = getType(method.getReturnType());
+        final ProbeType type = getType(method.getReturnType());
         if (type == null) {
             throw new IllegalArgumentException(format("@Probe method '%s.%s() has an unsupported return type'",
                     method.getDeclaringClass().getName(), method.getName()));
@@ -183,7 +183,7 @@ abstract class MethodProbe implements ProbeFunction {
     }
 
     static class LongMethodProbe<S> extends MethodProbe implements LongProbeFunction<S> {
-        LongMethodProbe(final Method method, final Probe probe, final ProbeUtils type, final SourceMetadata sourceMetadata) {
+        LongMethodProbe(final Method method, final Probe probe, final ProbeType type, final SourceMetadata sourceMetadata) {
             super(method, probe, type, sourceMetadata);
         }
 
@@ -215,7 +215,7 @@ abstract class MethodProbe implements ProbeFunction {
     }
 
     static class DoubleMethodProbe<S> extends MethodProbe implements DoubleProbeFunction<S> {
-        DoubleMethodProbe(final Method method, final Probe probe, final ProbeUtils type, final SourceMetadata sourceMetadata) {
+        DoubleMethodProbe(final Method method, final Probe probe, final ProbeType type, final SourceMetadata sourceMetadata) {
             super(method, probe, type, sourceMetadata);
         }
